Resolve Issue#1 (#2)
* Checks for file, if need be decompresses it

* Make sure the compressed log doesn't exist too

* Changed gzip to gunzip for decompression

* Compress succesful parsed log files

* Had square bracket too close

* Added flag to skip compression

* Fixed errors that I came across

* Removed echo that was for testing

* Improved reporting for an invalid argument

* Code Review Changes

+++ download.sh
@@ -3,6 +3,11 @@
 # Require $PROXYUSER, $PROXYPASS, $MYSITE; Optionally set $EZPFILESDIR
 source `dirname $0`/common.env
 
+function validate_url() {
+	wget -q --spider --load-cookies $SESSIONFILE $1
+	return $?
+}
+
 DAYSPRIOR=14
 if [[ "$1" =~ ^[0-9]+$ ]]
 then
@@ -32,13 +37,32 @@ fi
 for ((i=1; i<=$DAYSPRIOR; i++))
 do
 	TARGETDATE=`date "+%Y%m%d" -d "$CURDATE -$i days"`
-	if [[ ! -e $EZPFILESDIR/downloads/ezp${TARGETDATE}.log ]]
+	if [[ ! -e $EZPFILESDIR/downloads/ezp${TARGETDATE}.log && ! -e $EZPFILESDIR/downloads/ezp${TARGETDATE}.log.gz ]]
 	then
-		wget -q --load-cookies $SESSIONFILE -O ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log  "https://login.${MYSITE}.idm.oclc.org/loggedin/admlog/ezp${TARGETDATE}.log"
-		if [[ $? != 0 ]]
+		if validate_url "https://login.${MYSITE}.idm.oclc.org/loggedin/admlog/ezp${TARGETDATE}.log"
 		then
-			rm ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log
-			>&2 echo 'Failed to download ezp'$TARGETDATE', wget exited with '$?
+			wget -q --load-cookies $SESSIONFILE -O ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log  "https://login.${MYSITE}.idm.oclc.org/loggedin/admlog/ezp${TARGETDATE}.log"
+			if [[ $? != 0 ]]
+			then
+				rm ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log
+				>&2 echo 'Failed to download ezp'$TARGETDATE'.log, wget exited with '$?
+			fi
+		elif validate_url "https://login.${MYSITE}.idm.oclc.org/loggedin/admlog/ezp${TARGETDATE}.log.gz"
+		then
+			wget -q --load-cookies $SESSIONFILE -O ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log.gz  "https://login.${MYSITE}.idm.oclc.org/loggedin/admlog/ezp${TARGETDATE}.log.gz"
+			if [[ $? != 0 ]]
+			then
+				rm ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log.gz
+				>&2 echo 'Failed to download ezp'$TARGETDATE'.log.gz, wget exited with '$?
+			else 
+				gunzip ${EZPFILESDIR}/downloads/ezp${TARGETDATE}.log.gz
+				if [[ $? != 0 ]]
+				then
+					>&2 echo 'Failed to decompress ezp'$TARGETDATE'.log.gz, gunzip exited with '$?
+				fi
+			fi
+		else
+			>&2 echo 'Neither ezp'$TARGETDATE'.log nor ezp'$TARGETDATE'.log.gz could be found on the server.'
 		fi
 	fi
 done
